fix: handle debug mode failure better; move endpoint; target es2022
